Convert BloggerInterface to module blogger.py
